+++ chaospy/distributions/collection/mv_normal.py
@@ -1,109 +1,101 @@
 """Multivariate Normal Distribution."""
+import logging
 import numpy
 from scipy import special
 
 from .normal import normal
 
 from ..baseclass import Dist
+from ..operators import LocScaling
 
 
-class MvNormal(Dist):
-    """
+class MvNormal(LocScaling):
+    r"""
     Multivariate Normal Distribution.
 
     Args:
-        loc (float, Dist):
+        mu (float, numpy.ndarray):
             Mean vector
-        scale (float, Dist):
+        scale (float, numpy.ndarray):
             Covariance matrix or variance vector if scale is a 1-d vector.
+        rotation (Sequence[int]):
+            The order of which to resolve conditionals.
+            Defaults to `range(len(distribution))` which is the same as
+            `p(x0), p(x1|x0), p(x2|x0,x1), ...`.
 
     Examples:
-        >>> distribution = chaospy.MvNormal([1, 2], [[1, 0.6], [0.6, 1]])
-        >>> distribution
-        MvNormal(loc=[1.0, 2.0], scale=[[1.0, 0.6], [0.6, 1.0]])
+        >>> distribution = chaospy.MvNormal([10, 20, 30],
+        ...     [[1, 0.2, 0.3], [0.2, 2, 0.4], [0.3, 0.4, 1]], rotation=[1, 2, 0])
+        >>> distribution  # doctest: +NORMALIZE_WHITESPACE
+        MvNormal(mu=[10, 20, 30],
+                 sigma=[[1.0, 0.2, 0.3], [0.2, 2.0, 0.4], [0.3, 0.4, 1.0]])
+        >>> chaospy.E(distribution)
+        array([10., 20., 30.])
         >>> chaospy.Cov(distribution)
-        array([[1. , 0.6],
-               [0.6, 1. ]])
-        >>> mesh = numpy.meshgrid(*[numpy.linspace(0, 1, 5)[1:-1]]*2)
-        >>> distribution.inv(mesh).round(4)
-        array([[[0.3255, 1.    , 1.6745],
-                [0.3255, 1.    , 1.6745],
-                [0.3255, 1.    , 1.6745]],
-        <BLANKLINE>
-               [[1.0557, 1.4604, 1.8651],
-                [1.5953, 2.    , 2.4047],
-                [2.1349, 2.5396, 2.9443]]])
-        >>> distribution.fwd(distribution.inv(mesh)).round(4)
-        array([[[0.25, 0.5 , 0.75],
-                [0.25, 0.5 , 0.75],
-                [0.25, 0.5 , 0.75]],
-        <BLANKLINE>
-               [[0.25, 0.25, 0.25],
-                [0.5 , 0.5 , 0.5 ],
-                [0.75, 0.75, 0.75]]])
-        >>> distribution.pdf(distribution.inv(mesh)).round(4)
-        array([[0.0991, 0.146 , 0.1452],
-               [0.1634, 0.1989, 0.1634],
-               [0.1452, 0.146 , 0.0991]])
+        array([[1. , 0.2, 0.3],
+               [0.2, 2. , 0.4],
+               [0.3, 0.4, 1. ]])
+        >>> mesh = numpy.mgrid[:2, :2, :2].reshape(3, -1)*.5+.1
+        >>> mesh
+        array([[0.1, 0.1, 0.1, 0.1, 0.6, 0.6, 0.6, 0.6],
+               [0.1, 0.1, 0.6, 0.6, 0.1, 0.1, 0.6, 0.6],
+               [0.1, 0.6, 0.1, 0.6, 0.1, 0.6, 0.1, 0.6]])
+        >>> mapped_samples = distribution.inv(mesh)
+        >>> mapped_samples.round(2)
+        array([[ 8.39,  8.85,  8.39,  8.85,  9.86, 10.32,  9.86, 10.32],
+               [17.64, 18.26, 19.72, 20.34, 17.77, 18.39, 19.85, 20.47],
+               [28.72, 30.25, 28.72, 30.25, 28.72, 30.25, 28.72, 30.25]])
+        >>> numpy.allclose(distribution.fwd(mapped_samples), mesh)
+        True
+        >>> distribution.pdf(mapped_samples).round(4)
+        array([0.0042, 0.0092, 0.0092, 0.0203, 0.0092, 0.0203, 0.0203, 0.0446])
         >>> distribution.sample(4).round(4)
-        array([[ 1.395 , -0.2003,  2.6476,  0.9553],
-               [ 3.1476,  0.6411,  1.5946,  1.7647]])
-        >>> distribution.mom((1, 2)).round(4)
-        7.4
+        array([[10.1583,  9.1555, 11.3267, 10.1527],
+               [21.2826, 19.2191, 17.4524, 19.9038],
+               [29.2714, 31.0016, 29.1834, 30.651 ]])
+
     """
 
-    def __init__(self, loc=[0, 0], scale=[[1, .5], [.5, 1]]):
-        loc = numpy.asfarray(loc)
-        scale = numpy.asfarray(scale)
-        assert len(loc) == len(scale)
-        self._repr = {"loc": loc.tolist(), "scale": scale.tolist()}
+    def __init__(
+            self,
+            mu,
+            sigma=None,
+            rotation=None,
+            scale=None,
+    ):
+        logger = logging.getLogger(__name__)
+        assert scale is None or sigma is None, (
+            "Parameters 'scale' and 'sigma' can not be provided at the same time")
+        self._repr = {"mu": numpy.array(mu).tolist()}
+        if scale is not None:
+            logger.warning("Argument `scale` is to be deprecated. "
+                           "Use argument `sigma=scale**2` instead.")
+            sigma = scale**2
+            self._repr["scale"] = numpy.array(scale).tolist()
+        elif sigma is not None:
+            self._repr["sigma"] = numpy.array(sigma).tolist()
 
-        C = numpy.linalg.cholesky(scale)
-        Ci = numpy.linalg.inv(C)
-        Dist.__init__(self, C=C, Ci=Ci, loc=loc)
+        LocScaling.__init__(
+            self, dist=normal(), mean=mu, covariance=sigma, rotation=rotation)
 
-    def _cdf(self, x, C, Ci, loc):
-        return special.ndtr(numpy.dot(Ci, (x.T-loc.T).T))
-
-    def _ppf(self, q, C, Ci, loc):
-        return (numpy.dot(C, special.ndtri(q)).T+loc.T).T
-
-    def _pdf(self, x, C, Ci, loc):
-        det = numpy.linalg.det(numpy.dot(C,C.T))
-        x_ = numpy.dot(Ci.T, (x.T-loc.T).T)
-        out = numpy.ones(x.shape)
-        out[0] =  numpy.e**(-.5*numpy.sum(x_*x_, 0))/numpy.sqrt((2*numpy.pi)**len(Ci)*det)
-        return out
-
-    def _lower(self, C, Ci, loc):
-        return -7.5*numpy.sqrt(numpy.diag(numpy.dot(C, C.T)))+loc
-
-    def _upper(self, C, Ci, loc):
-        return 7.5*numpy.sqrt(numpy.diag(numpy.dot(C, C.T)))+loc
-
-    def _mom(self, k, C, Ci, loc):
-        scale = numpy.dot(C, C.T)
+    def _mom(self, k, cache):
         out = 0.
         for idx, kdx in enumerate(numpy.ndindex(*[_+1 for _ in k])):
             coef = numpy.prod(special.comb(k.T, kdx).T, 0)
-            diff = k.T - kdx
+            diff = k.T-kdx
             pos = diff >= 0
             diff = diff*pos
             pos = numpy.all(pos)
-            loc_ = numpy.prod(loc**diff)
+            location_ = numpy.prod(self.mean**diff)
 
-            out += pos*coef*loc_*isserlis_moment(tuple(kdx), scale)
+            out += pos*coef*location_*isserlis_moment(tuple(kdx), self.covariance)
 
         return float(out)
-
-    def __len__(self):
-        return len(self.prm["C"])
 
 
 def isserlis_moment(k, scale):
     """
-    Raw statistical moments centralized Normal distribution using Isserlis'
-    theorem.
+    Centralized statistical moments using Isserlis' theorem.
 
     Args:
         k (Tuple[int, ...]):
@@ -145,7 +137,6 @@ def isserlis_moment(k, scale):
     eye = numpy.eye(len(k), dtype=int)
     out = (k[idx]-1)*scale[idx, idx]*isserlis_moment(k-2*eye[idx], scale)
     for idj in range(idx+1, len(k)):
-        out += k[idj]*scale[idx, idj]*isserlis_moment(
-            k-eye[idx]-eye[idj], scale)
+        out += k[idj]*scale[idx, idj]*isserlis_moment(k-eye[idx]-eye[idj], scale)
 
     return float(out)