+++ chaospy/distributions/operators/__init__.py
@@ -5,8 +5,10 @@ from .matmul import Matmul
 from .negative import Neg, neg
 from .power import Pow
 
-from .joint import J, Iid
+from .joint import J
+from .iid import Iid
 
 from .trunkation import Trunc, Trunk, trunk
 
 from .logarithm import Log, Log10, Logn
+from .locscale import LocScaling