+++ chaospy/distributions/operators/joint.py
@@ -32,15 +32,6 @@ The created multivariate distribution behaves much like the univariate case::
      [ 0.4444  0.4444  0.5556  0.5556  0.7778  0.7778]]
     >>> print(distribution.mom([[2, 4, 6], [1, 2, 3]]))
     [0.5  1.   3.75]
-
-Constructing a multivariate probability distribution consisting of identical
-independent distributed marginals can be done using the
-:func:`~chaospy.distributions.operators.joint.Iid`. E.g.::
-
-    >>> X = chaospy.Normal()
-    >>> Y = chaospy.Iid(X, 4)
-    >>> print(Y.sample())
-    [ 0.39502989 -1.20032309  1.64760248 -0.04465437]
 """
 from copy import deepcopy
 
@@ -55,8 +46,10 @@ class J(Dist):
     Joint random variable generator.
 
     Args:
-        args: Dist
+        args (chaospy.Dist):
             Distribution to join together.
+        rotation (Optional[Sequence[int]]):
+            The order of how the joint should be evaluated.
     """
 
     @property
@@ -64,12 +57,21 @@ class J(Dist):
         """Determine if joint consist of only integers."""
         return all(dist.interpret_as_integer for dist in self.prm.values())
 
-    def __init__(self, *args):
+    def __init__(self, *args, **kwargs):
         args = [dist for arg in args
                 for dist in (arg if isinstance(arg, J) else [arg])]
         assert all(isinstance(dist, Dist) for dist in args)
         self.indices = [0] + numpy.cumsum([len(dist) for dist in args[:-1]]).tolist()
         self.inverse_map = {dist: idx for idx, dist in zip(self.indices, args)}
+        self._dependencies = [
+            deps.copy()
+            for dist in args
+            for deps in dist._dependencies
+        ]
+        if "rotation" in kwargs:
+            self._rotation = list(kwargs.pop("rotation"))
+        assert not kwargs, "'rotation' is the only allowed keyword."
+
         prm = {"_%03d" % idx: dist for idx, dist in zip(self.indices, args)}
         Dist.__init__(self, **prm)
 
@@ -100,8 +102,8 @@ class J(Dist):
         """
         Example:
             >>> dist = chaospy.J(chaospy.Uniform(), chaospy.Normal())
-            >>> dist.lower
-            array([ 0. , -7.5])
+            >>> dist.lower.round(4)
+            array([ 0.    , -6.3613])
             >>> d0 = chaospy.Uniform()
             >>> dist = chaospy.J(d0, d0+chaospy.Uniform())
             >>> dist.lower
@@ -119,8 +121,8 @@ class J(Dist):
         """
         Example:
             >>> dist = chaospy.J(chaospy.Uniform(), chaospy.Normal())
-            >>> dist.upper
-            array([1. , 7.5])
+            >>> dist.upper.round(4)
+            array([1.    , 6.3613])
             >>> d0 = chaospy.Uniform()
             >>> dist = chaospy.J(d0, d0+chaospy.Uniform())
             >>> dist.upper
@@ -230,10 +232,6 @@ class J(Dist):
                 dist, kloc[idx:idx+len(dist)], cache=cache)
             output.T[idx] = values.T
         return output
-
-
-    def __len__(self):
-        return sum(len(dist) for dist in self.inverse_map)
 
     def __str__(self):
         """